requirements.txt
# requirements.txt

streamlit
openai
matplotlib
opencv-python-headless
easyocr
cv2
PIL
python-dotenv